requirements.txt
azure-ai-documentintelligence==1.0.2
openai==1.97.1
streamlit==1.47.0
python-dotenv
pymupdf
python-docx==1.2.0
azure-identity